+++ proto/memcache/binary/pinger_test.go
@@ -25,10 +25,7 @@ func TestPingerPingEOF(t *testing.T) {
 	assert.NoError(t, err)
 
 	err = pinger.Ping()
-	assert.Error(t, err)
-
-	err = errors.Cause(err)
-	assert.Equal(t, bufio.ErrBufferFull, err)
+	assert.NoError(t, err)
 }
 
 func TestPingerPing100Ok(t *testing.T) {
@@ -41,8 +38,16 @@ func TestPingerPing100Ok(t *testing.T) {
 	}
 
 	err := pinger.Ping()
-	assert.Error(t, err)
-	_causeEqual(t, bufio.ErrBufferFull, err)
+	assert.NoError(t, err)
+}
+
+func TestPingerFlushErr(t *testing.T) {
+	conn := _createRepeatConn(pongBs, 100)
+	c := conn.Conn.(*mockConn)
+	c.err = errors.New("some error")
+	pinger := NewPinger(conn)
+	err := pinger.Ping()
+	assert.EqualError(t, err, "some error")
 }
 
 func TestPingerClosed(t *testing.T) {