+++ proto/memcache/pinger_test.go
@@ -3,8 +3,6 @@ package memcache
 import (
 	"testing"
 
-	"overlord/lib/bufio"
-
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 )
@@ -17,7 +15,7 @@ func TestPingerPingOk(t *testing.T) {
 	assert.NoError(t, err)
 }
 
-func TestPingerPingEOF(t *testing.T) {
+func TestPingerPingMore(t *testing.T) {
 	conn := _createConn(pongBytes)
 	pinger := NewPinger(conn)
 
@@ -25,10 +23,7 @@ func TestPingerPingEOF(t *testing.T) {
 	assert.NoError(t, err)
 
 	err = pinger.Ping()
-	assert.Error(t, err)
-
-	err = errors.Cause(err)
-	assert.Equal(t, bufio.ErrBufferFull, err)
+	assert.NoError(t, err)
 }
 
 func TestPingerPing100Ok(t *testing.T) {
@@ -41,8 +36,16 @@ func TestPingerPing100Ok(t *testing.T) {
 	}
 
 	err := pinger.Ping()
-	assert.Error(t, err)
-	_causeEqual(t, bufio.ErrBufferFull, err)
+	assert.NoError(t, err)
+}
+
+func TestPingerErr(t *testing.T) {
+	conn := _createRepeatConn(pongBytes, 100)
+	c := conn.Conn.(*mockConn)
+	c.err = errors.New("some error")
+	pinger := NewPinger(conn)
+	err := pinger.Ping()
+	assert.EqualError(t, err, "some error")
 }
 
 func TestPingerClosed(t *testing.T) {