+++ proto/redis/cluster/cluster.go
@@ -14,9 +14,6 @@ import (
 	"overlord/lib/log"
 	libnet "overlord/lib/net"
 	"overlord/proto"
-	"overlord/proto/redis"
-
-	"github.com/pkg/errors"
 )
 
 const (
@@ -40,8 +37,8 @@ const (
 	fakeClusterSlots = "" +
 		"*3\r\n" +
 		"*3\r\n:0\r\n:5460\r\n*2\r\n${IPLEN}\r\n{IP}\r\n:{PORT}\r\n" +
-		"*3\r\n:5460\r\n:10922\r\n*2\r\n${IPLEN}\r\n{IP}\r\n:{PORT}\r\n" +
-		"*3\r\n:10922\r\n:16383\r\n*2\r\n${IPLEN}\r\n{IP}\r\n:{PORT}\r\n"
+		"*3\r\n:5461\r\n:10922\r\n*2\r\n${IPLEN}\r\n{IP}\r\n:{PORT}\r\n" +
+		"*3\r\n:10923\r\n:16383\r\n*2\r\n${IPLEN}\r\n{IP}\r\n:{PORT}\r\n"
 )
 
 type cluster struct {
@@ -54,19 +51,15 @@ type cluster struct {
 	slotNode atomic.Value
 	action   chan struct{}
 
-	redts map[string]*redirect
-	rLock sync.Mutex
-
-	state int32
-
-	once sync.Once
-
 	fakeNodesBytes []byte
 	fakeSlotsBytes []byte
+	once           sync.Once
+
+	state int32
 }
 
-// NewExecutor new Executor.
-func NewExecutor(name, listen string, servers []string, conns int32, dto, rto, wto time.Duration, hashTag []byte) proto.Executor {
+// NewForwarder new proto Forwarder.
+func NewForwarder(name, listen string, servers []string, conns int32, dto, rto, wto time.Duration, hashTag []byte) proto.Forwarder {
 	c := &cluster{
 		name:    name,
 		servers: servers,
@@ -76,7 +69,6 @@ func NewExecutor(name, listen string, servers []string, conns int32, dto, rto, w
 		wto:     wto,
 		hashTag: hashTag,
 		action:  make(chan struct{}),
-		redts:   make(map[string]*redirect),
 	}
 	if !c.tryFetch() {
 		panic("redis cluster all seed nodes fail to fetch")
@@ -86,30 +78,21 @@ func NewExecutor(name, listen string, servers []string, conns int32, dto, rto, w
 	return c
 }
 
-func (c *cluster) Execute(mba *proto.MsgBatchAllocator, msgs []*proto.Message) error {
+func (c *cluster) Forward(msgs []*proto.Message) error {
 	if state := atomic.LoadInt32(&c.state); state == closed {
 		return ErrClusterClosed
 	}
 	for _, m := range msgs {
 		if m.IsBatch() {
 			for _, subm := range m.Batch() {
-				addr := c.getAddr(subm.Request().Key())
-				mba.AddMsg(addr, subm)
+				ncp := c.getPipe(subm.Request().Key())
+				ncp.Push(subm)
 			}
 		} else {
-			addr := c.getAddr(m.Request().Key())
-			mba.AddMsg(addr, m)
+			ncp := c.getPipe(m.Request().Key())
+			ncp.Push(m)
 		}
 	}
-	for addr, mb := range mba.MsgBatchs() {
-		if mb.Count() > 0 {
-			// WaitGroup add one MsgBatch!!!
-			mba.Add(1) // NOTE: important!!! for wait all MsgBatch done!!!
-			sn := c.slotNode.Load().(*slotNode)
-			sn.nodeChan[addr].push(mb)
-		}
-	}
-	mba.Wait()
 	return nil
 }
 
@@ -120,56 +103,13 @@ func (c *cluster) Close() error {
 	return nil
 }
 
-func (c *cluster) process(addr string) *batchChan {
-	nbc := newBatchChan(c.conns)
-	for i := int32(0); i < c.conns; i++ {
-		ncCh := &ncChan{nc: newNodeConn(c, addr), stop: make(chan struct{})}
-		nbc.add(i, ncCh)
-		go c.processIO(c.name, addr, nbc.ch, ncCh)
-	}
-	return nbc
-}
-
-func (c *cluster) processIO(name, addr string, ch chan *proto.MsgBatch, ncCh *ncChan) {
-	var (
-		nc  = ncCh.nc
-		err error
-	)
-	for {
-		if err != nil {
-			nc = newNodeConn(c, addr)
-			// TODO: there may be chaos with unexcepted closed of cluster.
-			ncCh.nc = nc
-			// fire try fetch action
-			select {
-			case c.action <- struct{}{}:
-			default:
-			}
-		}
-		mb, ok := <-ch
-		if !ok {
-			close(ncCh.stop) // NOTE: close stop, make sure nc closed concurrent security!!!
-			return
-		}
-		if err = nc.WriteBatch(mb); err != nil {
-			err = errors.Wrap(err, "Cluster batch write")
-			mb.DoneWithError(name, addr, err)
-			continue
-		}
-		if err := nc.ReadBatch(mb); err != nil {
-			err = errors.Wrap(err, "Cluster batch read")
-			mb.DoneWithError(name, addr, err)
-			continue
-		}
-		mb.Done(name, addr)
-	}
-}
-
-func (c *cluster) getAddr(key []byte) (addr string) {
+func (c *cluster) getPipe(key []byte) (ncp *proto.NodeConnPipe) {
 	realKey := c.trimHashTag(key)
 	crc := hashkit.Crc16(realKey) & musk
 	sn := c.slotNode.Load().(*slotNode)
-	return sn.nSlots.slots[crc]
+	addr := sn.nSlots.slots[crc]
+	ncp = sn.nodePipe[addr]
+	return
 }
 
 func (c *cluster) trimHashTag(key []byte) []byte {
@@ -208,75 +148,73 @@ func (c *cluster) tryFetch() bool {
 		f := newFetcher(conn)
 		nSlots, err := f.fetch()
 		if err != nil {
-			log.Errorf("fail to fetch due to %s", err)
+			if log.V(1) {
+				log.Errorf("Redis Cluster fail to fetch error:%v", err)
+			}
 			continue
 		}
 		c.initSlotNode(nSlots)
+		if log.V(4) {
+			log.Info("Redis Cluster try fetch success")
+		}
 		return true
 	}
-
-	log.Error("redis cluster all seed nodes fail to fetch")
+	if log.V(1) {
+		log.Error("Redis Cluster all seed nodes fail to fetch")
+	}
 	return false
 }
 
 func (c *cluster) initSlotNode(nSlots *nodeSlots) {
 	osn, ok := c.slotNode.Load().(*slotNode) // old slotNode
-	onc := map[string]*batchChan{}           // old nodeConn
+	oncp := map[string]*proto.NodeConnPipe{} // old nodeConn
 	if ok && osn != nil {
-		for addr, bc := range osn.nodeChan {
-			onc[addr] = bc // COPY
+		for addr, ncp := range osn.nodePipe {
+			oncp[addr] = ncp // COPY
 		}
 	}
 	sn := &slotNode{nSlots: nSlots}
-	sn.nodeChan = make(map[string]*batchChan)
+	sn.nodePipe = make(map[string]*proto.NodeConnPipe)
 	masters := nSlots.getMasters()
 	for _, addr := range masters {
-		bc, ok := onc[addr]
+		ncp, ok := oncp[addr]
 		if !ok {
-			bc = c.process(addr)
+			toAddr := addr // NOTE: avoid closure
+			ncp = proto.NewNodeConnPipe(c.conns, func() proto.NodeConn {
+				return newNodeConn(c, toAddr)
+			})
+			go c.pipeEvent(ncp.ErrorEvent())
+			if log.V(4) {
+				log.Infof("Redis Cluster renew slot node and add addr:%s", toAddr)
+			}
 		} else {
-			delete(onc, addr)
+			delete(oncp, addr)
+			if log.V(5) {
+				log.Infof("Redis Cluster renew slot node addr:%s", addr)
+			}
 		}
-		sn.nodeChan[addr] = bc
+		sn.nodePipe[addr] = ncp
 	}
 	c.servers = masters
 	c.slotNode.Store(sn)
-	for _, bc := range onc {
-		bc.close()
+	for addr, ncp := range oncp {
+		ncp.Close()
+		if log.V(4) {
+			log.Infof("Redis Cluster renew slot node and close addr:%s", addr)
+		}
 	}
 }
 
-func (c *cluster) getRedirectNodeConn(addr string) (r *redirect) {
-	c.rLock.Lock()
-	r, ok := c.redts[addr]
-	if ok {
-		c.rLock.Unlock()
-		return
-	}
-	rnc := redis.NewNodeConn(c.name, addr, c.dto, c.rto, c.wto).(*redis.NodeConn)
-	r = &redirect{nc: rnc}
-	c.redts[addr] = r
-	c.rLock.Unlock()
-	return
-}
-
-func (c *cluster) closeRedirectNodeConn(addr string, isAsk bool) {
-	if isAsk {
-		return
-	}
-	c.rLock.Lock()
-	r, ok := c.redts[addr]
-	if ok {
-		r.lock.Lock()
-		_ = r.nc.Close()
-		r.nc = nil
-		r.lock.Unlock() // FIXME(felix): when NodeConn have nc pointer in func redirectProcess
-		delete(c.redts, addr)
-	}
-	c.rLock.Unlock()
-	select {
-	case c.action <- struct{}{}:
-	default:
+func (c *cluster) pipeEvent(errCh <-chan error) {
+	for {
+		err, ok := <-errCh
+		if !ok {
+			return
+		}
+		if log.V(2) {
+			log.Errorf("Redis Cluster NodeConnPipe action error:%v", err)
+		}
+		c.action <- struct{}{}
 	}
 }
 
@@ -318,49 +256,7 @@ func (c *cluster) fake(listen string) {
 	})
 }
 
-type redirect struct {
-	nc   *redis.NodeConn
-	lock sync.Mutex
-}
-
 type slotNode struct {
 	nSlots   *nodeSlots
-	nodeChan map[string]*batchChan
-}
-
-type batchChan struct {
-	ch    chan *proto.MsgBatch
-	ncChs map[int32]*ncChan
-	state int32
-}
-
-func newBatchChan(n int32) *batchChan {
-	return &batchChan{ch: make(chan *proto.MsgBatch, n*1024), ncChs: make(map[int32]*ncChan)}
-}
-
-func (c *batchChan) push(m *proto.MsgBatch) {
-	if state := atomic.LoadInt32(&c.state); state == closed {
-		return
-	}
-	c.ch <- m
-}
-
-func (c *batchChan) add(i int32, ncCh *ncChan) {
-	c.ncChs[i] = ncCh
-}
-
-func (c *batchChan) close() {
-	if !atomic.CompareAndSwapInt32(&c.state, opening, closed) {
-		return
-	}
-	close(c.ch)
-	for _, ncCh := range c.ncChs {
-		<-ncCh.stop // NOTE: wait stop closed, make sure nc closed concurrent security!!!
-		_ = ncCh.nc.Close()
-	}
-}
-
-type ncChan struct {
-	nc   proto.NodeConn
-	stop chan struct{}
+	nodePipe map[string]*proto.NodeConnPipe
 }