+++ proto/redis/pinger_test.go
@@ -3,8 +3,7 @@ package redis
 import (
 	"testing"
 
-	"overlord/lib/bufio"
-
+	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -20,7 +19,7 @@ func TestPingerClosed(t *testing.T) {
 	p := NewPinger(conn)
 	assert.NoError(t, p.Close())
 	err := p.Ping()
-	assert.Equal(t, ErrPingClosed, err)
+	assert.Equal(t, ErrPingClosed, errors.Cause(err))
 	assert.NoError(t, p.Close())
 }
 
@@ -28,10 +27,19 @@ func TestPingerWrongResp(t *testing.T) {
 	conn := _createConn([]byte("-Error: iam more than 7 bytes\r\n"))
 	p := NewPinger(conn)
 	err := p.Ping()
-	assert.Equal(t, bufio.ErrBufferFull, err)
+	assert.Equal(t, ErrBadPong, errors.Cause(err))
 
 	conn = _createConn([]byte("-Err\r\n"))
 	p = NewPinger(conn)
 	err = p.Ping()
-	assert.Equal(t, ErrBadPong, err)
+	assert.Equal(t, ErrBadPong, errors.Cause(err))
+}
+
+func TestPingerPingErr(t *testing.T) {
+	conn := _createConn(pongBytes)
+	c := conn.Conn.(*mockConn)
+	c.err = errors.New("some error")
+	p := NewPinger(conn)
+	err := p.Ping()
+	assert.EqualError(t, err, "some error")
 }