+++ proxy/proxy.go
@@ -2,6 +2,7 @@ package proxy
 
 import (
 	errs "errors"
+	"net"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -27,8 +28,8 @@ type Proxy struct {
 	c   *Config
 	ccs []*ClusterConfig
 
-	executors map[string]proto.Executor
-	once      sync.Once
+	forwarders map[string]proto.Forwarder
+	once       sync.Once
 
 	conns int32
 
@@ -51,27 +52,31 @@ func New(c *Config) (p *Proxy, err error) {
 func (p *Proxy) Serve(ccs []*ClusterConfig) {
 	p.once.Do(func() {
 		p.ccs = ccs
-		p.executors = map[string]proto.Executor{}
+		p.forwarders = map[string]proto.Forwarder{}
 		if len(ccs) == 0 {
 			log.Warnf("overlord will never listen on any port due to cluster is not specified")
 		}
 		for _, cc := range ccs {
-			go p.serve(cc)
+			p.serve(cc)
 		}
 	})
 }
 
 func (p *Proxy) serve(cc *ClusterConfig) {
-	executor := NewExecutor(cc)
+	forwarder := NewForwarder(cc)
 	p.lock.Lock()
-	p.executors[cc.Name] = executor
+	p.forwarders[cc.Name] = forwarder
 	p.lock.Unlock()
 	// listen
 	l, err := Listen(cc.ListenProto, cc.ListenAddr)
 	if err != nil {
 		panic(err)
 	}
-	log.Infof("overlord proxy cluster[%s] addr(%s) already listened", cc.Name, cc.ListenAddr)
+	log.Infof("overlord proxy cluster[%s] addr(%s) start listening", cc.Name, cc.ListenAddr)
+	go p.accept(cc, l, forwarder)
+}
+
+func (p *Proxy) accept(cc *ClusterConfig, l net.Listener, forwarder proto.Forwarder) {
 	for {
 		conn, err := l.Accept()
 		if err != nil {
@@ -106,7 +111,7 @@ func (p *Proxy) serve(cc *ClusterConfig) {
 				continue
 			}
 		}
-		NewHandler(p, cc, conn, executor).Handle()
+		NewHandler(p, cc, conn, forwarder).Handle()
 	}
 }
 
@@ -117,8 +122,8 @@ func (p *Proxy) Close() error {
 	if p.closed {
 		return nil
 	}
-	for _, executor := range p.executors {
-		executor.Close()
+	for _, forwarder := range p.forwarders {
+		forwarder.Close()
 	}
 	return nil
 }