+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Overlord
 
+## Version 1.5.2
+1. max redirects 5.
+
+## Version 1.5.1
+1. reset sub message only in nedd.
+
+## Version 1.5.0
+1. refactor message pipeline.
+2. non-persistent connection for cluster redirect.
+
 ## Version 1.4.0
 1. add redis cluster support.